backend/requirements.txt
absl-py
acres
annotated-doc
annotated-types
anyio
certifi
charset-normalizer
ci-info
click
colorama
configobj
configparser
distro
etelemetry
fastapi
filelock
groq
h11
httpcore
httplib2
httptools
httpx
idna
jiter
joblib
looseversion
lxml
networkx
nibabel
nipype
nltk
numpy
openai
packaging
pandas
pathlib
portalocker
prov
puremagic
pydantic
pydantic_core
pydot
PyMuPDF
pyparsing
python-dateutil
python-dotenv
python-multipart
pytz
pyxnat
PyYAML
rdflib
regex
requests
rouge_score
sacrebleu
scipy
simplejson
six
sniffio
starlette
tabulate
tqdm
traits
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
websockets
